tout ok

+++ ex00/BitcoinExchange.hpp
@@ -21,7 +21,7 @@ class BitcoinExchange {
         void print( const char* input );
 
     private:
-        std::map<std::string, float> _data;
+        std::map<std::string, double> _data;
         
 };
